Created a basic dashboard
The dashboard contains 4 graphs with the top left graph to display user input data. Tried my best to match the sketch for the dashboard. The data that the graph is using is a default data that the API uses to demo for now. The graph is from the API library ReCharts.